+++ gatsby-config.js
@@ -1,12 +1,4 @@
 const path = require(`path`);
-const {
-  NODE_ENV,
-  URL: NETLIFY_SITE_URL = 'https://jacobdcastro.com',
-  DEPLOY_PRIME_URL: NETLIFY_DEPLOY_URL = NETLIFY_SITE_URL,
-  CONTEXT: NETLIFY_ENV = NODE_ENV,
-} = process.env;
-const isNetlifyProduction = NETLIFY_ENV === 'production';
-const siteUrl = isNetlifyProduction ? NETLIFY_SITE_URL : NETLIFY_DEPLOY_URL;
 
 module.exports = {
   pathPrefix: `/blog`,
@@ -119,22 +111,13 @@ module.exports = {
       options: {
         host: 'https://jacobdcastro.com',
         sitemap: 'https://jacobdcastro.com/sitemap.xml',
-        resolveEnv: () => NETLIFY_ENV,
+        resolveEnv: () => process.env.GATSBY_ENV,
         env: {
+          development: {
+            policy: [{ userAgent: '*', disallow: ['/'] }],
+          },
           production: {
-            policy: [
-              { userAgent: '*', allow: '/', disallow: ['/form-success'] },
-            ],
-          },
-          'branch-deploy': {
-            policy: [{ userAgent: '*', disallow: ['/'] }],
-            sitemap: null,
-            host: null,
-          },
-          'deploy-preview': {
-            policy: [{ userAgent: '*', disallow: ['/'] }],
-            sitemap: null,
-            host: null,
+            policy: [{ userAgent: '*', allow: '/', disallow: '/form-success' }],
           },
         },
       },