+++ src/pages/blog.js
@@ -8,17 +8,19 @@ import BlogListing from '../components/index/BlogListing';
 
 const Blog = ({ path, data }) => {
   const seo = {
-    page: `blog`,
-    title: `Blog`,
-    description: `I like writing about things. Life updates, random epiphanies, cool lessons, etc. Find 'em here!`,
-    url: `https://jacobdcastro.com/blog`,
+    page: 'blog',
+    title: 'Blog',
+    description:
+      // eslint-disable-next-line quotes
+      "I like writing about things. Life updates, random epiphanies, cool lessons, etc. Find 'em here!",
+    url: 'https://jacobdcastro.com/blog',
     imgUrl: `${data.pageImg.publicURL}`,
     imgAlt:
       'jdcastro logo, twitter, instagram, facebook, github icons with @jacobdcastro username',
     breadcrumbs: [
       {
-        name: `Blog`,
-        path: `/blog`,
+        name: 'Blog',
+        path: '/blog',
       },
     ],
   };